Window pop up terminal added to both pages
A run code button has been added to each webpage where when clicked, it opens a pop up window. This pop up window will serve as a terminal showing the output of the code  in the ide when the code is run. It has been tested with text output and user text input and it does print out both in the popup window.